refactor: modulate http responses and json-formatting to shared-structure
