+++ src/shared/http/network.c
@@ -1,69 +1,26 @@
-#include "http.h"
-#include <stdio.h>
-#include <stdlib.h>
+#include "network.h"
 #include <string.h>
 
 #ifdef _WIN32
-    // Windows
     #define WIN32_LEAN_AND_MEAN
     #include <windows.h>
     #include <winsock2.h>
-    #include <ws2tcpip.h>
     #include <iphlpapi.h>
+    #include <ws2tcpip.h>
 #else
-    // Unix-like
-    #include <unistd.h>
     #include <ifaddrs.h>
     #include <netdb.h>
     #include <net/if.h>
     #include <arpa/inet.h>
     #include <sys/socket.h>
     #include <sys/types.h>
+    #include <unistd.h>
+    #include <netinet/in.h>
 #endif
-
-void sendHttpResponse(int clientConnection, int statusCode, const char* statusMessage, 
-        const char* contentType, const char* body, const char* connection) {
-    int requiredSize = snprintf(NULL, 0, 
-        "HTTP/1.1 %d %s\r\n"
-        "Access-Control-Allow-Origin: *\r\n"
-        "Content-Type: %s\r\n"
-        "Content-Length: %d\r\n"
-        "Connection: %s\r\n"
-        "\r\n"
-        "%s", 
-        statusCode, statusMessage, contentType, (int)strlen(body), connection, body) + 1;
-        
-    char* response = (char*)malloc(requiredSize);
-    if (!response) return;
-    
-    snprintf(response, requiredSize,
-        "HTTP/1.1 %d %s\r\n"
-        "Access-Control-Allow-Origin: *\r\n"
-        "Content-Type: %s\r\n"
-        "Content-Length: %d\r\n"
-        "Connection: %s\r\n"
-        "\r\n"
-        "%s", 
-        statusCode, statusMessage, contentType, (int)strlen(body), connection, body);
-    
-    #ifdef _WIN32
-        send(clientConnection, response, (int)strlen(response), 0);
-    #else
-        write(clientConnection, response, strlen(response));
-    #endif
-    
-    free(response);
-}
-
-void sendErrorResponse(int clientConnection, int statusCode, 
-        const char* statusMessage, const char* body) {
-    sendHttpResponse(clientConnection, statusCode, statusMessage, 
-        "text/plain", body, "close");
-}
 
 int getLocalIP(char* ip, size_t ipSize) {
     #ifdef _WIN32
-        // Windows 
+        // Windows implementation
         PIP_ADAPTER_INFO pAdapterInfo;
         PIP_ADAPTER_INFO pAdapter = NULL;
         DWORD dwRetVal = 0;
@@ -91,7 +48,12 @@ int getLocalIP(char* ip, size_t ipSize) {
                     while (pIpAddr) {
                         if (strcmp(pIpAddr->IpAddress.String, "0.0.0.0") != 0 &&
                             strcmp(pIpAddr->IpAddress.String, "127.0.0.1") != 0) {
-                            strncpy_s(ip, ipSize, pIpAddr->IpAddress.String, _TRUNCATE);
+                            #ifdef _MSC_VER
+                                strncpy_s(ip, ipSize, pIpAddr->IpAddress.String, _TRUNCATE);
+                            #else
+                                strncpy(ip, pIpAddr->IpAddress.String, ipSize - 1);
+                                ip[ipSize - 1] = '\0';
+                            #endif
                             free(pAdapterInfo);
                             return 0;
                         }
@@ -105,7 +67,7 @@ int getLocalIP(char* ip, size_t ipSize) {
         free(pAdapterInfo);
         return 1;
     #else
-        // Unix-like
+        // Unix-like implementation
         struct ifaddrs *ifaddr, *ifa;
         int family, s;
         char host[NI_MAXHOST];
